#10 Divide swagger config from main.ts
