+++ server/src/main.ts
@@ -1,18 +1,10 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import {DocumentBuilder, SwaggerBaseConfig, SwaggerModule} from '@nestjs/swagger';
+import {setSwagger} from './config/swagger.config';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  // Swagger config
-  const options: SwaggerBaseConfig = new DocumentBuilder()
-      .setTitle('Beaver Server')
-      .setDescription('Beaver API description')
-      .addTag('Beaver')
-      .setVersion('1.0')
-      .build();
-  const document = SwaggerModule.createDocument(app, options);
-  SwaggerModule.setup('api', app, document);
+  setSwagger(app);
   await app.listen(3000);
 }
 bootstrap();
